hello-spring-boot-autoconfigure/src/main/java/com.lun/hello/spring/boot/autoconfigure/HelloAutoConfiguration.java
package com.lun.hello.spring.boot.autoconfigure;

import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
import org.springframework.boot.context.properties.EnableConfigurationProperties;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;

/**
 * @author lun
 * @create 2024/2/4 - 13:55
 */
@Configuration
@EnableConfigurationProperties(HelloProperties.class)
public class HelloAutoConfiguration {
    private final HelloProperties properties;

    public HelloAutoConfiguration(HelloProperties properties) {
        this.properties = properties;
    }
    @Bean
    @ConditionalOnMissingBean(User.class)
    public User user(){
        return new User(properties.getName());
    }
}
